backend/requirements.txt
# TaskLine Backend Dependencies

# Core Framework
fastapi>=0.114.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0

# Database
sqlmodel>=0.0.21,<1.0.0
alembic>=1.12.1,<2.0.0
supabase>=2.0.0,<3.0.0
psycopg2-binary>=2.9.9

# Authentication & Security
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt==4.3.0
pyjwt>=2.8.0,<3.0.0

# Settings & Validation
pydantic>=2.0
pydantic-settings>=2.2.1,<3.0.0
email-validator>=2.1.0,<3.0.0

# HTTP Client (for testing)
httpx>=0.25.1,<1.0.0

# Python multipart (for form data)
python-multipart>=0.0.7,<1.0.0
